postgres_to_es/requirements.txt
black==22.10.0
flake8==5.0.4
psutil==5.9.4
psycopg2-binary==2.9.4
pydantic==1.10.2
pyflakes==2.5.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0

